practice/main/requirements.txt
asgiref==3.5.2
azure-common==1.1.28
azure-core==1.25.1
azure-functions==1.12.0
azure-identity==1.11.0
azure-keyvault-secrets==4.6.0
azure-search-documents==11.4.0b3
azure-storage-blob==12.13.1
azure-storage-queue==12.5.0
backend==0.2.4.1
certifi==2022.6.15.2
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Django==4.0.7
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-storages==1.13.1
djangorestframework==3.14.0
idna==3.4
isodate==0.6.1
msal==1.19.0
msal-extensions==1.0.0
msrest==0.7.1
mssql-django==1.1.3
oauthlib==3.2.1
Pillow==9.2.0
portalocker==2.5.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.5.0
pyodbc==4.0.34
pytz==2022.2.1
pywin32==304
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tk==0.1.0
types-cryptography==3.3.23
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.12